Add a launcher '--dry-run' command line option.
Add a '--dry-run' option, which causes the Chapel launcher to print the
system launcher command line it would have used, including environment
settings, but not actually to run that system launcher.  In cases where
the launcher creates associated files such as workload manager batch
scripts and `expect` files, it also causes those files to be left for
inspection instead of cleaned up, and the user is responsible for
removing them when they are no longer of interest.

+++ runtime/src/launch/pbs-aprun/launch-pbs-aprun.c
@@ -379,7 +379,7 @@ static void genQsubScript(int argc, char *argv[], int numLocales) {
 }
 
 static void chpl_launch_cleanup(void) {
-  if (!debug) {
+  if (!chpl_doDryRun() && !debug) {
     if (unlink(expectFilename)) {
       char msg[FILENAME_MAX + 35];
       snprintf(msg, FILENAME_MAX + 35, "Error removing temporary file '%s': %s",
